% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pop1.R
\docType{data}
\name{Pop1}
\alias{Pop1}
\title{Pop1 cat body length example data}
\format{
\subsection{\code{Pop1}}{

A data frame with 30 rows and 3 columns:
\describe{
\item{x}{Population number}
\item{Body_length}{Cat body length data}
\item{ID}{Identification number of cat individuals}
}
}
}
\usage{
Pop1
}
\description{
A radomly generated data set to play with rerollR.
}
\details{
Use the following code (package stats required):

set.seed(42) #allow to generate the exact same data than in the user manual

Pop1 = data.frame(x=1, "Body_length" = c(runif(30,min=60,max=70)), "ID" = trunc(seq(0,29)/3))

Pop2 = data.frame(x=2, "Body_length" = c(runif(30,min=60,max=85)), "ID" = trunc(seq(0,29)/3))
}
\keyword{datasets}
