% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/re_boot.R
\name{re_boot}
\alias{re_boot}
\title{Use bootstrapping to generate a data frame}
\usage{
re_boot(data = NULL, popm, ID, n, boot.n = 1)
}
\arguments{
\item{data}{(input: NULL or a data frame) The input must be a data frame
containing the population metric measurements (identified by the popm
argument) and an identifier (a vector or column header) allowing to identify
the objects/individuals composing the population (identified by the ID
argument). Data argument input must be NULL (default) if popm and ID inputs
are vectors (e.g. data = NULL, popm = df$result, ID = df$column_name) and
should be a data frame if popm and ID inputs are column headers of this data
frame (e.g. data = df, popm = 'result', ID = 'column_name').}

\item{popm}{(input: a column header from the data argument or a vector) The
vector or column must contain all replicated measurements of each
object/individual of the population. These data are the one that will be
iteratively bootstrapped. NA values within popm input are omitted. The
column or vector length of popm must match the one of ID and be ordered in
the exact same way.}

\item{ID}{(input: NULL, a column header from the data argument or a vector)
The vector or column must contain PERFECTLY MATCHING identifiers (e.g.
character strings, ID numbers) regarding popm input, to allow the function
to identify objects/individuals within the population. The column or vector
length must match the one of popm and be ordered in the exact same way.}

\item{n}{(input: a positive integer) Specifies the number of alternative
populations generated by the function and stored in the data frame in
output. A n >= 10 000 is generally advised for reliable subsequent roll_
testings, while n=100 or 1000 can be useful for quick preliminary testing.}

\item{boot.n}{(input: a positive integer) Specifies the number of random
sampling with replacement performed within popm data for each of the
object/individual and each generation of alternative population. With boot.n
= 1 (default) the function takes only one measurement per object/individual
for each generation of alternative population. At boot.n >= 2, each
object/individual of each generated population become the mean of boot.n
numbers of bootstrap sampling from the original object/individual
measurement replicates.}
}
\value{
Return a data frame containing all created alternative populations.
For this data frame nrow is equal to the number of unique individual/object
with popm replicated values, and ncol = n. Can serve as a input for
roll_ functions.
}
\description{
Takes a population of objects/individuals known through replicated
measurements of a given metric. The function makes random sampling with
replacement (bootstrapping) within each object/individual data in order to
produce alternative populations.
}
\examples{
# The data frame mentioned in the example (Pop1) is contained in
# rerollR attached data.

#Example of processing using re_boot

#For not aggregated data frame (the only possibility)
re_boot(data=Pop1, popm="Body_length", ID ="ID", n=1000, boot.n = 1)
#equivalent to
re_boot(popm=Pop1$Body_length, ID =Pop1$ID, n=1000, boot.n = 1)

#With averaging after boot.n = x bootstrapping per individual/object
re_boot(data=Pop1, popm="Body_length", ID ="ID", n=1000, boot.n = 10)
}
