% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rerollR-package.R
\docType{package}
\name{rerollR-package}
\alias{rerollR}
\alias{rerollR-package}
\title{rerollR: Indirect computing of uncertainties of measurements in statistical tests}
\description{
An indirect method for exploring the effect of uncertainties of measurements on statistical test results. From one population of data, re_ functions generate “alternative populations” using Gaussian random number generation or bootstrapping. Together, these “alternative populations” explore the space of intra-individual (or intra-object) uncertainties of measurements. From two sets of “alternative populations”, roll_ functions perform series oft-tests, Wilcoxon rank sum tests or randomization tests and provide metrics and graphic representations of resulting p-value distributions.
}
\author{
\strong{Maintainer}: Auguste Hassler \email{auguste.hassler@gmail.com} (\href{https://orcid.org/0000-0003-3004-4680}{ORCID})

}
\keyword{internal}
